Move MissingFunctionException and the missing_function definition to pyglet\libs\.__init__ so other modules can use them without importing graphics.
