+++ pyglet/libs/__init__.py
@@ -17,3 +17,28 @@ Simple docstrings with minimal formatting are best because:
 2. The docstrings will be used to debug complex platform issues
 3. IDEs mangle formatting in any hover tooltips while debugging
 """
+
+from typing import NoReturn
+from collections.abc import Callable, Sequence
+
+
+class MissingFunctionException(Exception):  # noqa: N818
+    def __init__(self, name: str, requires: str | None = None,
+                 suggestions: Sequence[str] | None = None) -> None:
+        msg = f'{name} is not exported by the available OpenGL driver.'
+        if requires:
+            msg += f'  {requires} is required for this functionality.'
+        if suggestions:
+            msg += '  Consider alternative(s) {}.'.format(', '.join(suggestions))
+        Exception.__init__(self, msg)
+
+
+def missing_function(name: str, requires: str | None = None,
+                     suggestions: Sequence[str] | None = None) -> Callable:
+    def MissingFunction(*_args, **_kwargs) -> NoReturn:  # noqa: ANN002, ANN003, N802
+        raise MissingFunctionException(name, requires, suggestions)
+
+    return MissingFunction
+
+
+__all__ = ['MissingFunctionException', 'missing_function']
