+++ pyglet/libs/darwin/agl.py
@@ -1,7 +1,9 @@
 """Wrapper for /System/Library/Frameworks/AGL.framework/Headers/agl.h
 
-Generated by tools/gengl.py.
+Generated by tools/gen_opengl/gengl.py.
 Do not modify this file.
+
+Note: This is no longer generated by gengl as there is no XML for AGL.
 """
 
 from ctypes import Structure, POINTER, c_int, c_long, c_ubyte, c_ulong
