+++ pyglet/libs/win32/lib_wgl.py
@@ -5,7 +5,8 @@ from ctypes import CFUNCTYPE, POINTER, c_char_p, c_int, cast
 from typing import Any, Callable, Sequence
 
 import pyglet
-from pyglet.graphics.api.gl.lib import decorate_function, missing_function
+from pyglet.graphics.api.gl.lib import decorate_function
+from pyglet.libs import missing_function
 from pyglet.util import asbytes
 
 __all__ = ['link_GL', 'link_GL_proxy', 'link_WGL', 'link_WGL_proxy']
